WL-2316 Clarify privacy settings in profile2
Currently the privacy option is called Everyone but it is actually only
available to logged in users. This updates the `en` language strings to
reflect this.

Used non-WebLearn specific language so it can be shared.

Also noticed while doing this that profile2 picks up `en_US` files in
favour of `en_GB`.